# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Web Dev Blog Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stats/posts:
    get:
      operationId: getPostStats
      tags: [stats]
      summary: Get per-post view statistics, sorted by views
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Post statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostStat"
...
  /profile:
    get:
      operationId: getDeveloperProfile
      tags: [profile]
      summary: Get published developer profile (public)
      responses:
        "200":
          description: Developer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProfile"
        "404":
          description: Profile not found or not published
    patch:
      operationId: updateDeveloperProfile
      tags: [profile]
      summary: Update developer profile (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeveloperProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProfile"
        "401":
          description: Unauthorized
  /profile/manage:
    get:
      operationId: getDeveloperProfileManage
      tags: [profile]
      summary: Get developer profile for admin editing
      responses:
        "200":
          description: Developer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeveloperProfile"
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    PostStat:
      type: object
      required: [id, title, slug, views, status, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        views: { type: integer }
        status: { type: string }
        categoryName: { type: ["string", "null"] }
        createdAt: { type: string }
...
    ProfileWorkExperience:
      type: object
      required: [title, company, period, bullets]
      properties:
        title: { type: string }
        company: { type: string }
        period: { type: string }
        bullets:
          type: array
          items: { type: string }
    ProfileEducation:
      type: object
      required: [degree, institution, period]
      properties:
        degree: { type: string }
        institution: { type: string }
        period: { type: string }
    ProfileProject:
      type: object
      required: [name, description]
      properties:
        name: { type: string }
        description: { type: string }
    ProfileSkillGroup:
      type: object
      required: [category, items]
      properties:
        category: { type: string }
        items: { type: string }
    ProfileLanguage:
      type: object
      required: [name, level]
      properties:
        name: { type: string }
        level: { type: string }
    DeveloperProfile:
      type: object
      required: [id, name, status, workExperience, education, projects, technicalSkills, languages, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        headline: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        portfolioUrl: { type: ["string", "null"] }
        aboutMe: { type: ["string", "null"] }
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/ProfileWorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/ProfileEducation"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProfileProject"
        technicalSkills:
          type: array
          items:
            $ref: "#/components/schemas/ProfileSkillGroup"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLanguage"
        status: { type: string, enum: [draft, published] }
        updatedAt: { type: string }
    DeveloperProfileUpdate:
      type: object
      properties:
        name: { type: string }
        headline: { type: string }
        phone: { type: string }
        email: { type: string }
        location: { type: string }
        portfolioUrl: { type: string }
        aboutMe: { type: string }
        workExperience:
          type: array
          items:
            $ref: "#/components/schemas/ProfileWorkExperience"
        education:
          type: array
          items:
            $ref: "#/components/schemas/ProfileEducation"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProfileProject"
        technicalSkills:
          type: array
          items:
            $ref: "#/components/schemas/ProfileSkillGroup"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/ProfileLanguage"
        status: { type: string, enum: [draft, published] }